ufscar-sd/requirements.txt
pyzmq>=25.0.0
msgpack>=1.0.5
pyyaml>=6.0
pytest>=7.4.0
